refact: add generic io exception
